Move traits out of detail namespace
