+++ cpp/doc/DEVELOPER_GUIDE.md
@@ -362,7 +362,7 @@ An example function is helpful.
 template <class LonLatItA,
           class LonLatItB,
           class OutputIt,
-          class T        = typename detail::iterator_vec_base_type<LonLatItA>>
+          class T        = typename cuspatial::iterator_vec_base_type<LonLatItA>>
 OutputIt haversine_distance(LonLatItA a_lonlat_first,
                             LonLatItA a_lonlat_last,
                             LonLatItB b_lonlat_first,
